+++ packages/wonder-blocks-cell/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @khanacademy/wonder-blocks-cell
 
+## 3.3.3
+
+### Patch Changes
+
+-   Updated dependencies [874081aa]
+-   Updated dependencies [874081aa]
+-   Updated dependencies [a9bf603a]
+    -   @khanacademy/wonder-blocks-tokens@1.2.0
+    -   @khanacademy/wonder-blocks-clickable@4.1.3
+    -   @khanacademy/wonder-blocks-layout@2.0.30
+
 ## 3.3.2
 
 ### Patch Changes