+++ packages/wonder-blocks-grid/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @khanacademy/wonder-blocks-grid
 
+## 2.0.31
+
+### Patch Changes
+
+-   Updated dependencies [874081aa]
+-   Updated dependencies [874081aa]
+-   Updated dependencies [a9bf603a]
+    -   @khanacademy/wonder-blocks-tokens@1.2.0
+    -   @khanacademy/wonder-blocks-layout@2.0.30
+
 ## 2.0.30
 
 ### Patch Changes