+++ packages/wonder-blocks-tokens/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @khanacademy/wonder-blocks-tokens
 
+## 1.2.0
+
+### Minor Changes
+
+-   874081aa: Add fade and mix functions
+
+### Patch Changes
+
+-   874081aa: Remove wonder-blocks-color dependency in favor of wonder-blocks-tokens
+-   a9bf603a: Move spacing tokens to wb-tokens
+
 ## 1.1.0
 
 ### Minor Changes